Update pack 6
>Added Event listeners in PlayController
>Added movings in Ship class

+++ Project Space - New Live/modules/GameObjects/GameObject.cs
@@ -16,7 +16,7 @@ namespace Project_Space___New_Live.modules.GameObjects
         /// <summary>
         /// масса объекта
         /// </summary>
-        protected int mass;
+        protected float mass;
 
         /// <summary>
         /// отображение объекта
